Delete Customer from System implemented

+++ Lakeshore Market/com/online/lakeshoremarket/domain/CustomerDomain.java
@@ -29,10 +29,65 @@ public class CustomerDomain {
 		return addressID;
 	}
 
-	public void deleteCustomer(int custID) {
+	public boolean deleteCustomer(int custID) {
+		boolean isCustomerDeleted = false ;
 		custDao = new CustomerDAO();
-//		custDao.deleteCustomerAddress();
-		custDao.deleteCustomer(custID);
+		int billingAddressID = 0;
+		int shippingAddressID = 0;
+		int funcReturn = 0;
+		
+		
+		
+		/*billingAddressID = getCustomerBillingAddress(custID);
+		shippingAddressID = getCustomerShippingAddress(custID);
+		if(0 != billingAddressID && 0!= shippingAddressID){
+			int billingAddressDeleted = 0;
+			int shippingAddressDeleted = 0;
+			billingAddressDeleted = custDao.deleteCustomerAddress(billingAddressID);
+			shippingAddressDeleted = custDao.deleteCustomerAddress(shippingAddressID);
+			if(0 != billingAddressDeleted && 0 != shippingAddressDeleted){
+				funcReturn = custDao.deleteCustomer(custID);
+			}
+		}*/
+		
+		billingAddressID = getCustomerBillingAddress(custID);
+		shippingAddressID = getCustomerShippingAddress(custID);
+		if(0 != billingAddressID && 0 != shippingAddressID){
+			funcReturn = custDao.deleteCustomer(custID);
+			if(0 != funcReturn){
+				int billingAddressDeleted = 0;
+				int shippingAddressDeleted = 0;
+				billingAddressDeleted = custDao.deleteCustomerAddress(billingAddressID);
+				shippingAddressDeleted = custDao.deleteCustomerAddress(shippingAddressID);
+				if(0 != billingAddressDeleted && 0 != shippingAddressDeleted){
+					isCustomerDeleted = true;
+				}
+			}
+		}
+		
+		
+		
+//		isCustomerDeleted = (funcReturn == 0) ? false : true;
+		return isCustomerDeleted;
+	}
+	
+	public void deleteCustomerAddress(int custID){
+		custDao = new CustomerDAO();
+		custDao.deleteCustomerAddress(custID);
+	}
+	
+	public int getCustomerBillingAddress(int custID){
+		int billingAddressID = 0;
+		custDao = new CustomerDAO();
+		billingAddressID = custDao.getCustomerBillingAddress(custID);
+		return billingAddressID;
+	}
+	
+	public int getCustomerShippingAddress(int custID){
+		int shippingAddressID = 0;
+		custDao = new CustomerDAO();
+		shippingAddressID = custDao.getCustomerShippingAddress(custID);
+		return shippingAddressID;
 	}
 
 }
